+++ scripts/release_siridb.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
-rm -r ~/workspace/dbtest/dbpath0/benchmark
-xterm -e ~/workspace/siridb-server_union/Release/siridb-server -c  ~/workspace/dbtest/siridb0.conf &
+if [ -d ~/workspace/dbtest/dbpath0/benchmark ]; then
+  rm -r ~/workspace/dbtest/dbpath0/benchmark
+fi
+
+xterm -e ~/workspace/siridb-server/Release/siridb-server -c  ~/workspace/dbtest/siridb0.conf &
